+++ benches/src/blend.rs
@@ -6,10 +6,12 @@ fn fill_tiny_skia(blend_mode: tiny_skia::BlendMode, bencher: &mut Bencher) {
     let mut paint1 = Paint::default();
     paint1.set_color_rgba8(50, 127, 150, 200);
     paint1.blend_mode = BlendMode::SourceOver;
+    paint1.anti_alias = false;
 
     let mut paint2 = Paint::default();
     paint2.set_color_rgba8(220, 140, 75, 180);
     paint2.blend_mode = blend_mode; // <-- variable
+    paint2.anti_alias = false;
 
     let path1 = {
         let mut pb = PathBuilder::new();